Initial ShaTuTor refactor, move code to new 'Objectives' class
