+++ ShaTuApp/src/main/java/edu/regis/shatu/objectives/AddBits.java
@@ -1,0 +1,319 @@
+package edu.regis.shatu.objectives;
+
+import java.math.BigInteger;
+import java.util.Random;
+
+import edu.regis.shatu.err.NonRecoverableException;
+import edu.regis.shatu.model.Hint;
+import edu.regis.shatu.model.KnowledgeComponentKind;
+import edu.regis.shatu.model.Step;
+import edu.regis.shatu.model.StepCompletion;
+import edu.regis.shatu.model.StepCompletionReply;
+import edu.regis.shatu.model.Student;
+import edu.regis.shatu.model.StudentModelFieldKind;
+import edu.regis.shatu.model.Task;
+import edu.regis.shatu.model.TutoringSession;
+import edu.regis.shatu.model.aol.Assessment;
+import edu.regis.shatu.model.aol.AssessmentLevel;
+import edu.regis.shatu.model.aol.BitOpExample;
+import edu.regis.shatu.model.aol.BitOpStep;
+import edu.regis.shatu.model.aol.PendingStep;
+import edu.regis.shatu.model.aol.PendingTask;
+import edu.regis.shatu.model.aol.ProblemType;
+import edu.regis.shatu.model.aol.StepSubType;
+import edu.regis.shatu.model.aol.TaskKind;
+import edu.regis.shatu.model.aol.Timeout;
+import edu.regis.shatu.svc.ServiceFactory;
+import edu.regis.shatu.svc.StudentModelSvc;
+import edu.regis.shatu.svc.TutorReply;
+
+public class AddBits extends Objective {
+
+    private static final int MAX_BITS_SIZE = 32;
+
+    public AddBits(Student student) {
+        super(student);
+    }
+
+    /**
+     * Handles client requests for a new add bits example.
+     *
+     * Generates two operands with pre-size n bits. These are added together
+     * using modulo-preSize arithmetic. For example, with preSize 8 bits:</br>
+     * Operand1: 11100111:231 </br>
+     * Operand2: 10101111:175 </br>
+     * Intermediate: 110010110:406 </br>
+     * Result: 10010110:150 (e.g., 406 % 256)
+     *
+     * @param session  the current tutoring session.
+     * @param jsonData a BitOpExample encoded object
+     * @return a TutorReply
+     */
+    @Override
+    public TutorReply example(TutoringSession session, String jsonData) {
+        Random rnd = new Random();
+
+        BitOpExample example = gson.fromJson(jsonData, BitOpExample.class);
+
+        int size = example.getPreSize(); // number of bits
+
+        if (size == 0) {
+            // ToDo: The tutor should generate the string length and timeout
+            // based on the the current student model.
+            size = rnd.nextInt(MAX_BITS_SIZE - 1) + 1;
+            example.setTimeOut(600);
+
+        } else if (size > MAX_BITS_SIZE) {
+            // The student is requesting practice for a specific string length.
+            size = MAX_BITS_SIZE;
+            example.setTimeOut(0);
+        }
+
+        example.setPreSize(size);
+
+        int maxOperandVal = (int) Math.pow(2.0d, size) - 1; // e.g., 2^8 - 1 = 255
+
+        long operand1 = rnd.nextLong((maxOperandVal - 1) + 1);
+        example.setOperand1Val(operand1);
+
+        long operand2 = rnd.nextLong((maxOperandVal - 1) + 1);
+        example.setOperand2Val(operand2);
+
+        // This is Mod size arithmetic
+        example.setPostSize(size);
+        long result = ((long) operand1) + ((long) operand2);
+
+        if (result > maxOperandVal) {
+            result = result % (maxOperandVal + 1);
+        }
+
+        example.setResultVal(result);
+
+        BitOpStep subStep = new BitOpStep();
+        subStep.setExample(example);
+        // ToDo: multistep should be determined by the student model.
+        subStep.setMultiStep(rnd.nextBoolean());
+
+        Step step = new Step(1, 0, StepSubType.ADD_BITS);
+
+        // ToDo: fix timeouts
+        Timeout timeout = new Timeout("Complete Step", 0, ":No-Op", "Exceed time");
+        step.setTimeout(timeout);
+
+        step.setData(gson.toJson(subStep));
+
+        Task task = new Task();
+        task.setKind(TaskKind.PROBLEM);
+        task.setType(ProblemType.ADD_BITS);
+        task.setDescription("addition modulo 2^256 the bits in the two operands");
+        task.addStep(step);
+
+        // Update the assessment data and save it to the database.
+        int dbId = KnowledgeComponentKind.fromString("Add Bits").dbId();
+        Assessment assessment = studentModel.findAssessment(dbId);
+        assessment.incrementExposures();
+
+        try {
+            StudentModelSvc modelSvc = ServiceFactory.findStudentModelSvc();
+            modelSvc.updateAssessment(studentModel, assessment, StudentModelFieldKind.ATTEMPTS);
+
+            PendingStep pendingStep = new PendingStep(step);
+            pendingStep.setCurrentHintIndex(0);
+            pendingStep.setNotifyTutor(true);
+            pendingStep.setIsCompleted(false);
+
+            PendingTask pendingTask = new PendingTask(task);
+            pendingTask.setCurrentStep(pendingStep);
+
+            TutorReply reply = new TutorReply(":Success");
+            reply.setData(gson.toJson(pendingTask));
+
+            return reply;
+
+        } catch (NonRecoverableException ex) {
+            return createError("Unknown error", ex);
+        }
+    }
+
+    @Override
+    public TutorReply hint(StepCompletion completion) {
+        System.out.println("Tutor hintAddBits");
+
+        StepCompletionReply stepReply = new StepCompletionReply();
+
+        stepReply.setIsCorrect(false);
+        stepReply.setIsRepeatStep(true);
+        stepReply.setIsNewStep(false);
+        stepReply.setIsNewTask(false);
+        stepReply.setIsNextStep(false);
+
+        Hint hintOne = new Hint();
+        hintOne.setId(0);
+        hintOne.setText("Add the bits as if they were unsigned integers, discarding any overflow beyond 32 bits");
+
+        Step step = completion.getStep();
+        step.addHint(hintOne);
+
+        step.setSubType(StepSubType.REQUEST_HINT);
+        Timeout timeout = new Timeout("Complete Step", 0, ":No-Op", "Exceed time");
+        step.setTimeout(timeout);
+        step.setData(gson.toJson(stepReply));
+
+        PendingStep pendingStep = new PendingStep(step);
+        pendingStep.setCurrentHintIndex(0);
+        pendingStep.setNotifyTutor(true);
+        pendingStep.setIsCompleted(false);
+
+        TutorReply reply = new TutorReply(":Success");
+        reply.setData(gson.toJson(pendingStep));
+
+        // Update the assessment data and save it to the database.
+        int dbId = KnowledgeComponentKind.fromString("Add Bits").dbId();
+        Assessment assessment = studentModel.findAssessment(dbId);
+        assessment.incrementHints();
+
+        try {
+            StudentModelSvc modelSvc = ServiceFactory.findStudentModelSvc();
+            modelSvc.updateAssessment(studentModel, assessment, StudentModelFieldKind.HINTS);
+
+        } catch (NonRecoverableException ex) {
+            return createError("Unknown error", ex);
+        }
+
+        return reply;
+    }
+
+    @Override
+    public TutorReply completeStep(StepCompletion completion) {
+        System.out.println("Tutor completeAddBitsStep");
+
+        BitOpStep example = gson.fromJson(completion.getData(), BitOpStep.class);
+
+        String operand1 = example.getExample().getOperand1();
+        String operand2 = example.getExample().getOperand2();
+        String result = example.getExample().getResult();
+
+        int m = 8; // this will be changed
+
+        String expectedResult = addBitsFunction(operand1, operand2, m);
+
+        StepCompletionReply stepReply = new StepCompletionReply();
+        stepReply.setCorrectAnswer(expectedResult);
+        stepReply.setResponse(result);
+
+        if (expectedResult.equals(result)) {
+            stepReply.setIsCorrect(true);
+            stepReply.setIsRepeatStep(false);
+            stepReply.setIsNewStep(true);
+
+            // ToDo: Use the student model to figure out whether we want
+            // to give the student another practice problem of the same
+            // type or move on to an entirely different problem.
+            stepReply.setIsNewTask(true);
+
+            // ToDo: currently only one step in a task, so there isn't a next one???
+            stepReply.setIsNextStep(false);
+
+            // Update the assessment data and save it to the database.
+            int dbId = KnowledgeComponentKind.fromString("Add Bits").dbId();
+            Assessment assessment = studentModel.findAssessment(dbId);
+            assessment.incrementSuccessess();
+
+            int exposures = assessment.getExposures();
+            int successes = assessment.getSuccessess();
+
+            if (exposures > 0 && (double) successes / exposures > 0.6) {
+                stepReply.setIsNewTask(true);
+                System.out.println("%%%%%%%%%%% Next Task Recommended");
+                assessment.setAssessment(AssessmentLevel.COMPLETED);
+            } else {
+                stepReply.setIsNewTask(false);
+            }
+
+            try {
+                StudentModelSvc modelSvc = ServiceFactory.findStudentModelSvc();
+                modelSvc.updateAssessment(studentModel, assessment, StudentModelFieldKind.SUCCESSES);
+                modelSvc.updateAssessment(studentModel, assessment, StudentModelFieldKind.ASSESSMENT_LEVEL);
+
+            } catch (NonRecoverableException ex) {
+                return createError("Unknown error", ex);
+            }
+
+        } else {
+            stepReply.setIsCorrect(false);
+            stepReply.setIsRepeatStep(true);
+            stepReply.setIsNewStep(false);
+            stepReply.setIsNewTask(false);
+            stepReply.setIsNextStep(false);
+        }
+
+        Step step = new Step(1, 0, StepSubType.STEP_COMPLETION_REPLY);
+
+        // ToDo: fix timeouts
+        Timeout timeout = new Timeout("Complete Step", 0, ":No-Op", "Exceed time");
+        step.setTimeout(timeout);
+
+        step.setData(gson.toJson(stepReply));
+
+        Task task = new Task();
+        task.setKind(TaskKind.PROBLEM);
+        task.setType(ProblemType.STEP_COMPLETION_REPLY);
+        task.setDescription("Choose your next action");
+        task.addStep(step);
+
+        PendingStep pendingStep = new PendingStep(step);
+        pendingStep.setCurrentHintIndex(0);
+        pendingStep.setNotifyTutor(true);
+        pendingStep.setIsCompleted(false);
+
+        PendingTask pendingTask = new PendingTask(task);
+        pendingTask.setCurrentStep(pendingStep);
+
+        TutorReply reply = new TutorReply(":Success");
+
+        reply.setData(gson.toJson(pendingTask));
+
+        return reply;
+    }
+
+    /**
+     * Performs binary addition modulo 2^256
+     *
+     * @param operand1 The binary string for operand 1.
+     * @param operand2 The binary string for operand 2.
+     * @param m        The int for calculating the modulo
+     * @return The binary string result after adding two bits
+     */
+    private String addBitsFunction(String operand1, String operand2, int m) {
+
+        if (operand1 == null || operand1.isEmpty()) {
+            return "";
+        }
+        if (operand2 == null || operand2.isEmpty()) {
+            return "";
+        }
+
+        // Convert binary strings to BigIntegers
+        BigInteger num1 = new BigInteger(operand1, 2);
+        BigInteger num2 = new BigInteger(operand2, 2);
+
+        // Perform addition
+        BigInteger sum = num1.add(num2);
+
+        // Calculate the result modulo 2^256
+        BigInteger modulo = new BigInteger("2").pow(m);
+        BigInteger result = sum.mod(modulo);
+
+        // Convert the result back to a binary string
+        String resultBinary = result.toString(2);
+
+        // Ensure the binary string has 256 bits (pad with leading zeros if necessary)
+        while (resultBinary.length() < m) {
+            resultBinary = "0" + resultBinary;
+        }
+
+        System.out.println("Result : " + resultBinary);
+
+        return resultBinary;
+    }
+}
