+++ ShaTuApp/src/main/java/edu/regis/shatu/objectives/ChoiceFunction.java
@@ -1,0 +1,262 @@
+package edu.regis.shatu.objectives;
+
+import edu.regis.shatu.err.NonRecoverableException;
+import edu.regis.shatu.model.ChoiceFunctionStep;
+import edu.regis.shatu.model.Hint;
+import edu.regis.shatu.model.KnowledgeComponentKind;
+import edu.regis.shatu.model.Step;
+import edu.regis.shatu.model.StepCompletion;
+import edu.regis.shatu.model.StepCompletionReply;
+import edu.regis.shatu.model.Student;
+import edu.regis.shatu.model.StudentModelFieldKind;
+import edu.regis.shatu.model.Task;
+import edu.regis.shatu.model.TutoringSession;
+import edu.regis.shatu.model.aol.Assessment;
+import edu.regis.shatu.model.aol.AssessmentLevel;
+import edu.regis.shatu.model.aol.PendingStep;
+import edu.regis.shatu.model.aol.PendingTask;
+import edu.regis.shatu.model.aol.ProblemType;
+import edu.regis.shatu.model.aol.StepSubType;
+import edu.regis.shatu.model.aol.TaskKind;
+import edu.regis.shatu.model.aol.Timeout;
+import edu.regis.shatu.svc.ServiceFactory;
+import edu.regis.shatu.svc.StudentModelSvc;
+import edu.regis.shatu.svc.TutorReply;
+
+public class ChoiceFunction extends Objective {
+    public ChoiceFunction(Student student) {
+        super(student);
+    }
+
+    @Override
+    public TutorReply completeStep(StepCompletion completion) {
+        System.out.println("Tutor completeChoiceStep");
+
+        ChoiceFunctionStep example = gson.fromJson(completion.getData(), ChoiceFunctionStep.class);
+        String operand1 = example.getOperand1();
+        String operand2 = example.getOperand2();
+
+        String operand3 = example.getOperand3();
+        int bitLength = example.getBitLength();
+        String result = example.getResult();
+
+        String expectedResult = choiceFunction(operand1, operand2, operand3, bitLength);
+
+        System.out.println("Expected result: " + expectedResult); // ToDo debuggin
+
+        StepCompletionReply stepReply = new StepCompletionReply();
+        stepReply.setCorrectAnswer(expectedResult);
+        stepReply.setResponse(result);
+
+        if (expectedResult.equals(result)) {
+            stepReply.setIsCorrect(true);
+            stepReply.setIsRepeatStep(false);
+            stepReply.setIsNewStep(true);
+
+            // ToDo: currently only one step in a task, so there isn't a next one???
+            stepReply.setIsNextStep(false);
+
+            // Update the assessment data and save it to the database.
+            int dbId = KnowledgeComponentKind.fromString("Choice Function").dbId();
+            Assessment assessment = studentModel.findAssessment(dbId);
+            assessment.incrementSuccessess();
+
+            int exposures = assessment.getExposures();
+            int successes = assessment.getSuccessess();
+
+            if (exposures > 0 && (double) successes / exposures > 0.8) {
+                stepReply.setIsNewTask(true);
+                System.out.println("%%%%%%%%%%% Next Task Recommended");
+                assessment.setAssessment(AssessmentLevel.COMPLETED);
+            } else {
+                stepReply.setIsNewTask(false);
+            }
+
+            try {
+                StudentModelSvc modelSvc = ServiceFactory.findStudentModelSvc();
+                modelSvc.updateAssessment(studentModel, assessment, StudentModelFieldKind.SUCCESSES);
+                modelSvc.updateAssessment(studentModel, assessment, StudentModelFieldKind.ASSESSMENT_LEVEL);
+
+            } catch (NonRecoverableException ex) {
+                return createError("Unknown error", ex);
+            }
+
+        } else {
+            stepReply.setIsCorrect(false);
+            stepReply.setIsRepeatStep(true);
+            stepReply.setIsNewStep(false);
+            stepReply.setIsNewTask(false);
+            stepReply.setIsNextStep(false);
+        }
+
+        Step step = new Step(1, 0, StepSubType.STEP_COMPLETION_REPLY);
+
+        // ToDo: fix timeouts
+        Timeout timeout = new Timeout("Complete Step", 0, ":No-Op", "Exceed time");
+        step.setTimeout(timeout);
+        step.setData(gson.toJson(stepReply));
+
+        Task task = new Task();
+        task.setKind(TaskKind.PROBLEM);
+        task.setType(ProblemType.STEP_COMPLETION_REPLY);
+        task.setDescription("Choose your next action");
+        task.addStep(step);
+
+        PendingStep pendingStep = new PendingStep(step);
+        pendingStep.setCurrentHintIndex(0);
+        pendingStep.setNotifyTutor(true);
+        pendingStep.setIsCompleted(false);
+
+        PendingTask pendingTask = new PendingTask(task);
+        pendingTask.setCurrentStep(pendingStep);
+
+        TutorReply reply = new TutorReply(":Success");
+
+        reply.setData(gson.toJson(pendingTask));
+
+        return reply;
+    }
+
+    /**
+     * Handles client requests for a new choice function example.
+     *
+     * @return a TutorReply
+     */
+    @Override
+    public TutorReply example(TutoringSession session, String jsonData) {
+        System.out.println("newChoiceFunctionExample");
+        ChoiceFunctionStep substep = gson.fromJson(jsonData, ChoiceFunctionStep.class);
+
+        int bitLength = substep.getBitLength();
+
+        String operand1 = generateInputString(bitLength);
+        String operand2 = generateInputString(bitLength);
+        String operand3 = generateInputString(bitLength);
+
+        substep.setOperand1(operand1);
+        substep.setOperand2(operand2);
+        substep.setOperand3(operand3);
+
+        substep.setResult(choiceFunction(operand1, operand2, operand3, bitLength));
+
+        Step step = new Step(1, 0, StepSubType.CHOICE_FUNCTION);
+
+        // ToDo: fix timeouts
+        Timeout timeout = new Timeout("Complete Step", 0, ":No-Op", "Exceed time");
+        step.setTimeout(timeout);
+
+        step.setData(gson.toJson(substep));
+
+        Task task = new Task();
+        task.setKind(TaskKind.PROBLEM);
+        task.setType(ProblemType.CHOICE_FUNCTION);
+        task.setDescription("Compute the result of the choice function on the three operands");
+        task.addStep(step);
+
+        // Update the assessment data and save it to the database.
+        int dbId = KnowledgeComponentKind.fromString("Choice Function").dbId();
+        Assessment assessment = studentModel.findAssessment(dbId);
+        assessment.incrementExposures();
+
+        try {
+            StudentModelSvc modelSvc = ServiceFactory.findStudentModelSvc();
+            modelSvc.updateAssessment(studentModel, assessment, StudentModelFieldKind.ATTEMPTS);
+
+            PendingStep pendingStep = new PendingStep(step);
+            pendingStep.setCurrentHintIndex(0);
+            pendingStep.setNotifyTutor(true);
+            pendingStep.setIsCompleted(false);
+
+            PendingTask pendingTask = new PendingTask(task);
+            pendingTask.setCurrentStep(pendingStep);
+
+            TutorReply reply = new TutorReply(":Success");
+            reply.setData(gson.toJson(pendingTask));
+
+            return reply;
+
+        } catch (NonRecoverableException ex) {
+            return createError("Unknown error", ex);
+        }
+    }
+
+    /**
+     * Evaluates the choice function Ch(x, y, z).
+     *
+     * @param x Binary string representation of x.
+     * @param y Binary string representation of y.
+     * @param z Binary string representation of z.
+     * @return Binary string result of Ch(x, y, z).
+     */
+    private String choiceFunction(String x, String y, String z, int bitLength) {
+        // Convert the binary strings to integer values
+        String tempX = x.replaceAll("\\s", "");
+        String tempY = y.replaceAll("\\s", "");
+        String tempZ = z.replaceAll("\\s", "");
+
+        long intX = Long.parseLong(tempX, 2);
+        long intY = Long.parseLong(tempY, 2);
+        long intZ = Long.parseLong(tempZ, 2);
+
+        long xy = intX & intY;
+
+        long notX = ~intX & intZ;
+
+        long result = xy ^ notX;
+
+        // Convert the result back to binary string
+        String binaryResult = formatResult(result, bitLength);
+
+        return binaryResult;
+    }
+
+    @Override
+    public TutorReply hint(StepCompletion completion) {
+        System.out.println("Tutor hintChoiceFunction");
+
+        StepCompletionReply stepReply = new StepCompletionReply();
+
+        stepReply.setIsCorrect(false);
+        stepReply.setIsRepeatStep(true);
+        stepReply.setIsNewStep(false);
+        stepReply.setIsNewTask(false);
+        stepReply.setIsNextStep(false);
+
+        Hint hintOne = new Hint();
+        hintOne.setId(0);
+        hintOne.setText(
+                "The choice function selects bits from one input or another based on the value of the first input");
+
+        Step step = completion.getStep();
+        step.addHint(hintOne);
+
+        step.setSubType(StepSubType.REQUEST_HINT);
+        Timeout timeout = new Timeout("Complete Step", 0, ":No-Op", "Exceed time");
+        step.setTimeout(timeout);
+        step.setData(gson.toJson(stepReply));
+
+        PendingStep pendingStep = new PendingStep(step);
+        pendingStep.setCurrentHintIndex(0);
+        pendingStep.setNotifyTutor(true);
+        pendingStep.setIsCompleted(false);
+
+        TutorReply reply = new TutorReply(":Success");
+        reply.setData(gson.toJson(pendingStep));
+
+        // Update the assessment data and save it to the database.
+        int dbId = KnowledgeComponentKind.fromString("Choice Function").dbId();
+        Assessment assessment = studentModel.findAssessment(dbId);
+        assessment.incrementHints();
+
+        try {
+            StudentModelSvc modelSvc = ServiceFactory.findStudentModelSvc();
+            modelSvc.updateAssessment(studentModel, assessment, StudentModelFieldKind.HINTS);
+
+        } catch (NonRecoverableException ex) {
+            return createError("Unknown error", ex);
+        }
+
+        return reply;
+    }
+
+}
