+++ ShaTuApp/src/main/java/edu/regis/shatu/objectives/EncodeAscii.java
@@ -1,0 +1,273 @@
+package edu.regis.shatu.objectives;
+
+import edu.regis.shatu.err.NonRecoverableException;
+import edu.regis.shatu.model.EncodeAsciiStep;
+import edu.regis.shatu.model.Hint;
+import edu.regis.shatu.model.KnowledgeComponentKind;
+import edu.regis.shatu.model.Step;
+import edu.regis.shatu.model.StepCompletion;
+import edu.regis.shatu.model.StepCompletionReply;
+import edu.regis.shatu.model.Student;
+import edu.regis.shatu.model.StudentModelFieldKind;
+import edu.regis.shatu.model.Task;
+import edu.regis.shatu.model.TutoringSession;
+import edu.regis.shatu.model.aol.Assessment;
+import edu.regis.shatu.model.aol.AssessmentLevel;
+import edu.regis.shatu.model.aol.PendingStep;
+import edu.regis.shatu.model.aol.PendingTask;
+import edu.regis.shatu.model.aol.ProblemType;
+import edu.regis.shatu.model.aol.StepSubType;
+import edu.regis.shatu.model.aol.TaskKind;
+import edu.regis.shatu.model.aol.Timeout;
+import edu.regis.shatu.svc.ServiceFactory;
+import edu.regis.shatu.svc.StudentModelSvc;
+import edu.regis.shatu.svc.TutorReply;
+
+public class EncodeAscii extends Objective {
+    public EncodeAscii(Student student) {
+        super(student);
+    }
+
+    /**
+     * This method is called from the EncodeAsciiView when the check button is
+     * clicked, will check the users answer to the correct answer generated by
+     * the newEncodeAsciiStep method in this file.
+     *
+     * @param completion
+     *
+     * @return
+     */
+    @Override
+    public TutorReply completeStep(StepCompletion completion) {
+        // EncodeAsciiStep from the StepCompletion function in the Encode Ascii View
+        EncodeAsciiStep completedEncodeAsciiStep = gson.fromJson(completion.getData(), EncodeAsciiStep.class);
+
+        String userAnswer = completedEncodeAsciiStep.getUserAnswer(); // What the user submitted as the answer.
+        String correctAnswer = completedEncodeAsciiStep.getResult();
+
+        System.out.println("Correct Answer: " + correctAnswer); // Error checking
+        System.out.println("User Answer: " + userAnswer); // Error checking
+
+        StepCompletionReply stepReply = new StepCompletionReply();
+        stepReply.setCorrectAnswer(correctAnswer);
+        stepReply.setResponse(userAnswer);
+
+        if (userAnswer.equals(correctAnswer)) { // User was correct
+            System.out.println("Answer was correct, correct if branch taken."); // Error checking.
+            stepReply.setIsCorrect(true);
+            stepReply.setIsRepeatStep(false);
+            stepReply.setIsNewStep(true);
+            stepReply.setIsNewTask(true);
+            stepReply.setIsNextStep(false);
+
+            // Update the assessment data and save it to the database.
+            int dbId = KnowledgeComponentKind.fromString("ASCII Encode").dbId();
+            Assessment assessment = studentModel.findAssessment(dbId);
+            assessment.incrementSuccessess();
+
+            int exposures = assessment.getExposures();
+            int successes = assessment.getSuccessess();
+
+            if (exposures > 0 && (double) successes / exposures > 0.6) {
+                stepReply.setIsNewTask(true);
+                System.out.println("%%%%%%%%%%% Next Task Recommended");
+                assessment.setAssessment(AssessmentLevel.COMPLETED);
+            } else {
+                stepReply.setIsNewTask(false);
+            }
+
+            try {
+                StudentModelSvc modelSvc = ServiceFactory.findStudentModelSvc();
+                modelSvc.updateAssessment(studentModel, assessment, StudentModelFieldKind.SUCCESSES);
+                modelSvc.updateAssessment(studentModel, assessment, StudentModelFieldKind.ASSESSMENT_LEVEL);
+
+            } catch (NonRecoverableException ex) {
+                return createError("Unknown error", ex);
+            }
+
+        } else { // User was wrong
+            System.out.println("Answer was not correct, correct if branch taken."); // Error checking
+            stepReply.setIsCorrect(false);
+            stepReply.setIsRepeatStep(true);
+            stepReply.setIsNewStep(false);
+            stepReply.setIsNewTask(false);
+            stepReply.setIsNextStep(false);
+        }
+
+        Step step = new Step(1, 0, StepSubType.STEP_COMPLETION_REPLY);
+
+        // ToDo: fix timeouts
+        Timeout timeout = new Timeout("Complete Step", 0, ":No-Op", "Exceed time");
+        step.setTimeout(timeout);
+        step.setData(gson.toJson(stepReply));
+
+        Task task = new Task();
+        task.setKind(TaskKind.PROBLEM);
+        task.setType(ProblemType.STEP_COMPLETION_REPLY);
+        task.setDescription("Choose your next action");
+        task.addStep(step);
+
+        PendingStep pendingStep = new PendingStep(step);
+        pendingStep.setCurrentHintIndex(0);
+        pendingStep.setNotifyTutor(true);
+        pendingStep.setIsCompleted(false);
+
+        PendingTask pendingTask = new PendingTask(task);
+        pendingTask.setCurrentStep(pendingStep);
+
+        TutorReply reply = new TutorReply(":Success");
+
+        reply.setData(gson.toJson(pendingTask));
+
+        return reply;
+    }
+
+    /**
+     * Handles client requests for a new ASCII encode example.
+     *
+     * @return a TutorReply whose data contains a JSon EncodeAsciiExample
+     *         object.
+     */
+    @Override
+    public TutorReply example(TutoringSession session, String jsonData) {
+        System.out.println("Start tutor newEncodeAsciiexample"); // Error checking
+
+        EncodeAsciiStep newEncodeAscii = gson.fromJson(jsonData, EncodeAsciiStep.class);
+
+        if (newEncodeAscii.getQuestion().isEmpty() || newEncodeAscii.getQuestion() == null) {
+
+            System.out.println("Question was empty"); // Error checking
+
+            int messageLength = newEncodeAscii.getMessageLength();
+
+            String newQuestion = generateRandomString(messageLength); // Generates a random string to convert to binary
+
+            newEncodeAscii.setQuestion(newQuestion);
+
+            newEncodeAscii.setResult(toBinaryFunction(newQuestion));
+
+        } else {
+            newEncodeAscii.setResult(toBinaryFunction(newEncodeAscii.getQuestion()));
+        }
+
+        System.out.println(newEncodeAscii.getResult()); // Error checking
+
+        Step step = new Step(1, 0, StepSubType.ENCODE_ASCII);
+
+        // ToDo: fix timeouts
+        Timeout timeout = new Timeout("Complete Step", 0, ":No-Op", "Exceed time");
+        step.setTimeout(timeout);
+
+        step.setData(gson.toJson(newEncodeAscii));
+
+        Task task = new Task();
+        task.setKind(TaskKind.PROBLEM);
+        task.setType(ProblemType.ASCII_ENCODE);
+        task.setDescription("Convert the question to binary.");
+        task.addStep(step);
+
+        // Update the assessment data and save it to the database.
+        int dbId = KnowledgeComponentKind.fromString("ASCII Encode").dbId();
+        Assessment assessment = studentModel.findAssessment(dbId);
+        assessment.incrementExposures();
+
+        try {
+            StudentModelSvc modelSvc = ServiceFactory.findStudentModelSvc();
+            modelSvc.updateAssessment(studentModel, assessment, StudentModelFieldKind.ATTEMPTS);
+
+            PendingStep pendingStep = new PendingStep(step);
+            pendingStep.setCurrentHintIndex(0);
+            pendingStep.setNotifyTutor(true);
+            pendingStep.setIsCompleted(false);
+
+            PendingTask pendingTask = new PendingTask(task);
+            pendingTask.setCurrentStep(pendingStep);
+
+            TutorReply reply = new TutorReply(":Success");
+            reply.setData(gson.toJson(pendingTask));
+
+            return reply;
+
+        } catch (NonRecoverableException ex) {
+            return createError("Unknown error", ex);
+        }
+    }
+
+    @Override
+    public TutorReply hint(StepCompletion completion) {
+        System.out.println("Tutor hintEncode");
+
+        StepCompletionReply stepReply = new StepCompletionReply();
+
+        stepReply.setIsCorrect(false);
+        stepReply.setIsRepeatStep(true);
+        stepReply.setIsNewStep(false);
+        stepReply.setIsNewTask(false);
+        stepReply.setIsNextStep(false);
+
+        Hint hintOne = new Hint();
+        hintOne.setId(0);
+        hintOne.setText("Convert each character to its ASCII value");
+
+        Hint hintTwo = new Hint();
+        hintTwo.setId(1);
+        hintTwo.setText("Each character should be represented by 8 bits");
+
+        Step step = completion.getStep();
+        step.addHint(hintOne);
+        step.addHint(hintTwo);
+
+        step.setSubType(StepSubType.REQUEST_HINT);
+        Timeout timeout = new Timeout("Complete Step", 0, ":No-Op", "Exceed time");
+        step.setTimeout(timeout);
+        step.setData(gson.toJson(stepReply));
+
+        PendingStep pendingStep = new PendingStep(step);
+        pendingStep.setCurrentHintIndex(0);
+        pendingStep.setNotifyTutor(true);
+        pendingStep.setIsCompleted(false);
+
+        TutorReply reply = new TutorReply(":Success");
+        reply.setData(gson.toJson(pendingStep));
+
+        // Update the assessment data and save it to the database.
+        int dbId = KnowledgeComponentKind.fromString("ASCII Encode").dbId();
+        Assessment assessment = studentModel.findAssessment(dbId);
+        assessment.incrementHints();
+
+        try {
+            StudentModelSvc modelSvc = ServiceFactory.findStudentModelSvc();
+            modelSvc.updateAssessment(studentModel, assessment, StudentModelFieldKind.HINTS);
+
+        } catch (NonRecoverableException ex) {
+            return createError("Unknown error", ex);
+        }
+
+        return reply;
+    }
+
+    /**
+     * Function that can take a string and convert it to binary
+     *
+     * @param question
+     * @return
+     */
+    private String toBinaryFunction(String question) {
+        String answer;
+
+        char stringArray[] = question.toCharArray();
+
+        StringBuilder binary = new StringBuilder();
+
+        for (int i = 0; i < stringArray.length; i++) {
+            String binaryChar = String.format("%8s", Integer.toBinaryString(stringArray[i])).replaceAll(" ", "0");
+
+            binary.append(binaryChar);
+        }
+
+        answer = binary.toString();
+
+        return answer;
+    }
+
+}
